+++ tests/test_kttk/test_template_manager.py
@@ -2,6 +2,7 @@ import os
 import pytest
 
 from mock import patch
+from valideer import ValidationError
 
 from kttk import template_manager
 
@@ -147,11 +148,8 @@ def test_route_template():
 
 def test_validate_routes():
     # test valid
-    is_valid, reason = template_manager._validate_routes({'test': '123'})
-    assert is_valid
-    assert reason == ''
+    template_manager.route_schema.validate({'test': '123'})
 
     # test invalid
-    is_valid, reason = template_manager._validate_routes({'test': []})
-    assert is_valid == False
-    assert len(reason) > 0
+    with pytest.raises(ValidationError):
+        template_manager.route_schema.validate({'test': []})